support fp16 with llama

+++ mlora/lora_liner.py
@@ -29,11 +29,11 @@ class Lora(torch.nn.Module):
         self.scaling_ = alpha / r
 
     def forward(self, data: torch.Tensor) -> torch.Tensor:
-        data_ = F.dropout(data, self.dropout_)
+        data_ = F.dropout(data.type_as(self.lora_a_), self.dropout_)
         data_ @= self.lora_a_.transpose(0, 1)
         data_ @= self.lora_b_.transpose(0, 1)
         data_ *= self.scaling_
-        return data_
+        return data_.type_as(data)
 
 
 class Linear(torch.nn.Module):
@@ -46,10 +46,8 @@ class Linear(torch.nn.Module):
             self.device_ = device
 
         if not isinstance(weight, torch.nn.Linear):
-            import bitsandbytes
-            assert isinstance(weight,
-                              bitsandbytes.nn.Linear8bitLt) or isinstance(weight,
-                                                                          bitsandbytes.nn.Linear4bit), "error type."
+            assert isinstance(weight, bitsandbytes.nn.Linear8bitLt) or isinstance(
+                weight, bitsandbytes.nn.Linear4bit), "error type."
         else:
             weight.requires_grad_(False)
